pc - rename package from demonextjsspringtodoapp to courses
