+++ pom.xml
@@ -194,7 +194,7 @@
 					<archive>
 						<manifest>
 							<!-- full package name of class with the main you want to run -->
-							<mainClass>edu.ucsb.demonextjsspringtodoapp.DemoApplication</mainClass>
+							<mainClass>edu.ucsb.courses.DemoApplication</mainClass>
 
 						</manifest>
 					</archive>
