+++ src/main/java/edu/ucsb/demonextjsspringtodoapp/controllers/AppController.java
@@ -1,4 +1,4 @@
-package edu.ucsb.demonextjsspringtodoapp.controllers;
+package edu.ucsb.courses.controllers;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,8 +9,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
-import edu.ucsb.demonextjsspringtodoapp.services.Auth0Service;
-import edu.ucsb.demonextjsspringtodoapp.models.GoogleUserProfile;
+import edu.ucsb.courses.services.Auth0Service;
+import edu.ucsb.courses.models.GoogleUserProfile;
 
 @RestController
 public class AppController {
