+++ src/main/java/edu/ucsb/demonextjsspringtodoapp/repositories/AdminRepository.java
@@ -1,6 +1,6 @@
-package edu.ucsb.demonextjsspringtodoapp.repositories;
+package edu.ucsb.courses.repositories;
 
-import edu.ucsb.demonextjsspringtodoapp.entities.Admin;
+import edu.ucsb.courses.entities.Admin;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
