+++ src/main/java/edu/ucsb/demonextjsspringtodoapp/repositories/AppUserRepository.java
@@ -1,6 +1,6 @@
-package edu.ucsb.demonextjsspringtodoapp.repositories;
+package edu.ucsb.courses.repositories;
 
-import edu.ucsb.demonextjsspringtodoapp.entities.AppUser;
+import edu.ucsb.courses.entities.AppUser;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
