+++ src/test/java/edu/ucsb/courses/entities/TodoTests.java
@@ -1,4 +1,4 @@
-package edu.ucsb.demonextjsspringtodoapp.entities;
+package edu.ucsb.courses.entities;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
