+++ src/test/java/edu/ucsb/courses/services/Auth0ServiceTests.java
@@ -1,10 +1,10 @@
-package edu.ucsb.demonextjsspringtodoapp.services;
+package edu.ucsb.courses.services;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
-import edu.ucsb.demonextjsspringtodoapp.models.GoogleUserProfile;
+import edu.ucsb.courses.models.GoogleUserProfile;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
